Refactor Trip Conditions on Ascent Modal for upcoming improvements
